requirements.txt
aiohttp==3.7.4
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
Automat==20.2.0
certifi==2020.11.8
chardet==3.0.4
constantly==15.1.0
discord.py==1.6.0
docopt==0.6.2
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
multidict==5.1.0
osascript==2020.12.3
pipreqs==0.4.10
psutil==5.8.0
requests==2.25.1
runcmd==2020.12.3
schedule==1.0.0
six==1.15.0
Twisted==21.2.0
typing-extensions==3.7.4.3
urllib3==1.26.3
yarg==0.1.9
yarl==1.6.3
zope.interface==5.3.0
